# api/openapi.yaml
openapi: 3.1.0
info:
  title: AshML API
  description: Control-plane API for the AshML Kubernetes-native GPU ML platform.
    Generated from the route schemas — do not edit by hand.
  version: 0.1.0-dev
components:
  schemas:
    def-0:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Stable, machine-readable error code
            message:
              type: string
      title: Error
...
    def-6:
      type: object
      required:
        - id
        - name
        - project
        - state
        - priority
        - resources
        - created_at
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        project:
          type: string
        state:
          type: string
          enum:
            - CREATED
            - QUEUED
            - SCHEDULING
            - STARTING
            - RUNNING
            - FAILED
            - RETRYING
            - CANCELLING
            - SUCCEEDED
            - CANCELLED
        priority:
          type: string
          enum:
            - HIGH
            - MEDIUM
            - LOW
        resources:
          type: object
          properties:
            cpu:
              type: integer
            memory_bytes:
              type: integer
            gpu:
              type: integer
            gpu_memory_min_bytes:
              type: integer
        spec:
          type: object
          additionalProperties: true
        experiment:
          type:
            - "null"
            - object
          description: The reproducibility record this run belongs to, if any
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
        attempt:
          type: integer
        max_retries:
          type: integer
        k8s_job_name:
          type:
            - "null"
            - string
          description: Name of the Kubernetes Job running this attempt, once launched
        failure_reason:
          type:
            - "null"
            - string
        placement:
          type: object
          properties:
            node_id:
              type:
                - "null"
                - string
            reason:
              type:
                - "null"
                - string
        queued_at:
          type:
            - "null"
            - string
          format: date-time
        started_at:
          type:
            - "null"
            - string
          format: date-time
        finished_at:
          type:
            - "null"
            - string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type:
            - "null"
            - string
          format: date-time
      title: TrainingJob
    def-7:
      type: object
      required:
        - id
        - event_type
        - created_at
      properties:
        id:
          type: integer
        event_type:
          type: string
        from_state:
          type:
            - "null"
            - string
        to_state:
          type:
            - "null"
            - string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
      title: JobEvent
paths:
...
  /api/v1/jobs:
    post:
      summary: Submit a training job
      tags:
        - jobs
      description: Creates the job and admits it to the queue in one transaction. The
        executor claims it from there and runs it as a Kubernetes Job.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - project
                - name
                - spec
              additionalProperties: false
              properties:
                project:
                  type: string
                experiment:
                  type: string
                  format: uuid
                  description: Optional experiment to attribute this run to; must be in the same
                    project
                name:
                  type: string
                  minLength: 1
                  maxLength: 63
                  pattern: ^[a-z0-9]([a-z0-9-]*[a-z0-9])?$
                priority:
                  type: string
                  enum:
                    - HIGH
                    - MEDIUM
                    - LOW
                  default: MEDIUM
                max_retries:
                  type: integer
                  minimum: 0
                  maximum: 10
                  default: 0
                resources:
                  type: object
                  additionalProperties: false
                  properties:
                    cpu:
                      type: integer
                      minimum: 0
                      default: 1
                    memory_bytes:
                      type: integer
                      minimum: 0
                      default: 0
                    gpu:
                      type: integer
                      minimum: 0
                      default: 0
                    gpu_memory_min_bytes:
                      type: integer
                      minimum: 0
                      default: 0
                spec:
                  type: object
                  required:
                    - image
                  properties:
                    image:
                      type: string
                      minLength: 1
                    command:
                      type: array
                      items:
                        type: string
                    args:
                      type: array
                      items:
                        type: string
                    env:
                      type: object
                      additionalProperties:
                        type: string
                  additionalProperties: true
      responses:
        "201":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-6"
        "400":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
    get:
      summary: List training jobs
      tags:
        - jobs
      parameters:
        - schema:
            type: string
          in: query
          name: project
          required: false
        - schema:
            type: string
            enum:
              - CREATED
              - QUEUED
              - SCHEDULING
              - STARTING
              - RUNNING
              - FAILED
              - RETRYING
              - CANCELLING
              - SUCCEEDED
              - CANCELLED
          in: query
          name: state
          required: false
        - schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 50
          in: query
          name: limit
          required: false
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - jobs
                properties:
                  jobs:
                    type: array
                    items:
                      $ref: "#/components/schemas/def-6"
  /api/v1/jobs/{id}:
    get:
      summary: Get a job
      tags:
        - jobs
      parameters:
        - schema:
            type: string
            format: uuid
          in: path
          name: id
          required: true
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-6"
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
  /api/v1/jobs/{id}/events:
    get:
      summary: Get a job's event history
      tags:
        - jobs
      description: The append-only audit trail of everything that happened to this
        job, oldest first (spec §47).
      parameters:
        - schema:
            type: string
            format: uuid
          in: path
          name: id
          required: true
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - events
                properties:
                  events:
                    type: array
                    items:
                      $ref: "#/components/schemas/def-7"
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
  /api/v1/jobs/{id}/cancel:
    post:
      summary: Cancel a job
      tags:
        - jobs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type:
                - object
                - "null"
              additionalProperties: false
              properties:
                reason:
                  type: string
                  maxLength: 500
      parameters:
        - schema:
            type: string
            format: uuid
          in: path
          name: id
          required: true
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-6"
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
        "409":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
  /api/v1/jobs/{id}/logs:
    get:
      summary: Read a job's container logs
      tags:
        - jobs
      description: "Returns the training container's stdout/stderr as read from
        Kubernetes. Logs live in the cluster, not in AshML: once the Pod is
        garbage-collected they are gone, which is why the response says whether
        they are still available rather than returning an empty string."
      parameters:
        - schema:
            type: integer
            minimum: 1
            maximum: 10000
          in: query
          name: tail
          required: false
          description: Return only the most recent N lines
        - schema:
            type: boolean
            default: false
          in: query
          name: previous
          required: false
          description: Read the previous container instance instead of the current one —
            where the output of a crashed run is found
        - schema:
            type: string
            format: uuid
          in: path
          name: id
          required: true
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - job_id
                  - available
                  - logs
                properties:
                  job_id:
                    type: string
                    format: uuid
                  state:
                    type: string
                    enum:
                      - CREATED
                      - QUEUED
                      - SCHEDULING
                      - STARTING
                      - RUNNING
                      - FAILED
                      - RETRYING
                      - CANCELLING
                      - SUCCEEDED
                      - CANCELLED
                  k8s_job_name:
                    type:
                      - "null"
                      - string
                  available:
                    type: boolean
                    description: False when no Pod exists to read logs from
                  reason:
                    type:
                      - "null"
                      - string
                    description: Why logs are unavailable, when they are not
                  logs:
                    type: string
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
servers:
  - url: /
    description: Current host
...